+++ src/interfaces/IDelegatedParameters.ts
@@ -1,6 +1,7 @@
 import {IIteration} from './IIteration';
 
 export interface IDelegatedParameters {
+    wd?: string;
     parameters?: {[key: string]: any};
     iteration?: IIteration;
 }